pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "maldet"
dynamic = ["version"]
description = "Plug-and-play malware detector framework"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Po-Lin Lai", email = "islab.ai.tool@gmail.com" }]
requires-python = ">=3.12"
keywords = ["malware", "security", "machine-learning", "detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "pydantic>=2.6",
    "hydra-core>=1.3",
    "hydra-zen>=0.13",
    "typer>=0.12",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.1",
    "structlog>=24.1",
    "joblib>=1.3",
    "jinja2>=3.1",
    "omegaconf>=2.3",
    "tomli_w>=1.0",
]

[project.optional-dependencies]
lightning = [
    "lightning>=2.5",
    "torch>=2.2",
    "torchmetrics>=1.3",
]
mlflow = ["mlflow>=2.20"]
all = ["maldet[lightning,mlflow]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.10",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "pre-commit>=3.7",
]

[project.scripts]
maldet = "maldet.cli:app"

[project.urls]
Homepage = "https://github.com/bolin8017/maldet"
Documentation = "https://bolin8017.github.io/maldet/"
Repository = "https://github.com/bolin8017/maldet.git"
Issues = "https://github.com/bolin8017/maldet/issues"

[tool.hatch.version]
path = "src/maldet/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/maldet"]

[tool.hatch.build.targets.wheel.force-include]
"src/maldet/templates" = "maldet/templates"

[tool.hatch.build.targets.sdist]
include = ["src/", "tests/", "docs/", "pyproject.toml", "README.md", "LICENSE", "CHANGELOG.md", "mkdocs.yml"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "N", "PT"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["maldet"]

[[tool.mypy.overrides]]
module = ["lightning.*", "torch.*", "torchmetrics.*", "mlflow.*", "hydra.*", "joblib.*", "sklearn.*", "pandas.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "integration: end-to-end tests that may be slow",
    "gpu: requires CUDA",
]

[tool.coverage.run]
source = ["src/maldet"]
branch = true
omit = ["src/maldet/_version.py"]

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pragma: no cover",
]
